# openapi.yaml
openapi: 3.0.3
info:
  title: Unikol API
  description: >
    API REST de la plateforme SaaS de gestion scolaire Unikol.
    Toutes les routes (sauf /auth/*) exigent un JWT Bearer valide.
    Le tenant (SchoolId) est dérivé du claim `schoolId` du token, jamais d'un paramètre de requête.
    Référence fonctionnelle complète : docs/Volume_4_API_Design.md
  version: "1.1.0"
servers:
  - url: https://api.sama-ecole.sn/v1
    description: Production
  - url: https://staging-api.sama-ecole.sn/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Développement local
security:
  - bearerAuth: []
...
paths:
...
  /admin/platform/dashboard:
    get:
      tags: [Platform]
      summary: Agrégats plateforme, toutes écoles confondues (Super Admin uniquement)
      description: >-
        Le Super Admin n'a aucun SchoolId propre : une requête EF Core normale sur les tables tenant
        (schools/users/subscriptions/subscription_payments) ne renverrait donc rien (RLS, AGENTS.md
        règle #2). Lit la vue PostgreSQL `v_platform_dashboard_stats` (migration
        AddPlatformAdminViews), qui contourne l'isolation via `security_invoker = false` +
        OWNER sama_ecole — jamais en affaiblissant les policies existantes.
      responses:
        '200':
          description: Agrégats plateforme
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformDashboardStats' }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
  /admin/platform/activity:
    get:
      tags: [Platform]
      summary: Journal d'audit de toutes les écoles, paginé (Super Admin uniquement)
      description: >-
        Même principe que /admin/platform/dashboard, mais pour `audit_logs` : appelle la fonction
        PostgreSQL SECURITY DEFINER `get_global_audit_logs(limit, offset)` (migration
        AddPlatformAdminViews), qui contourne la RLS pour ce seul rôle de façon auditée et volontaire.
        `audit_logs` reste par ailleurs une table tenant normale (RLS + Global Query Filter) pour
        tous les autres appelants.
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedGlobalAuditLogs' }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
  /admin/platform/subscriptions:
    get:
      tags: [Platform]
      summary: Abonnements de toutes les écoles, avec dernier paiement confirmé (Super Admin uniquement)
      description: >-
        Même principe que /admin/platform/dashboard : lit la vue PostgreSQL `v_platform_subscriptions`
        (migration AddPlatformSubscriptionsAndImpersonation), qui contourne la RLS de
        `subscriptions`/`subscription_payments` via `security_invoker = false` + OWNER sama_ecole.
        Liste complète, non paginée (nombre d'écoles modeste), filtrage/recherche côté client comme
        GET /schools.
      responses:
        '200':
          description: Liste des abonnements
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PlatformSubscription' } }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
  /admin/platform/revenue-projection:
    get:
      tags: [Platform]
      summary: Projection de revenu sur 12 mois, basée sur les échéances réelles (Super Admin uniquement)
      description: >-
        Alimente le graphique de projection de l'Executive Dashboard. Pour chaque école à
        l'abonnement Active, projette ses échéances de renouvellement futures (ExpiresAt, puis +N
        fois la période de facturation de son dernier paiement confirmé) sur les 12 prochains mois et
        cumule le montant attendu par mois. Contrairement au MRR (équivalent mensuel lissé), reflète
        le cash réellement attendu mois par mois — un abonnement Yearly n'apparaît qu'au(x) mois de
        son échéance. Aucune hypothèse de croissance (nouvelles écoles) : voir le Simulateur de
        Croissance, purement côté client (ARPU × nombre d'écoles cible), pour ce scénario.
      responses:
        '200':
          description: 12 points mensuels, du mois courant inclus aux 11 mois suivants
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MonthlyRevenueProjection' } }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
  /admin/platform/subscriptions/{schoolId}/remind:
    post:
      tags: [Platform]
      summary: Envoyer un rappel de paiement au Directeur de l'établissement (Super Admin uniquement)
      description: >-
        Bouton « Relancer » de l'écran Abonnements & Facturation. N'écrit AUCUN paiement ni statut
        d'abonnement (AGENTS.md règle #11) : envoie un e-mail de rappel au Directeur actif de
        l'établissement, retrouvé via la fonction SECURITY DEFINER
        `auth_find_active_director_by_school`. Écrit une entrée d'audit sur l'école ciblée.
      parameters:
        - name: schoolId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Rappel envoyé }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
        '404': { description: Établissement introuvable }
        '409': { description: Aucun compte Directeur actif pour cet établissement }
  /admin/platform/schools/{schoolId}/impersonate:
    post:
      tags: [Platform]
      summary: Ouvrir une session Directeur pour l'établissement ciblé (Super Admin uniquement)
      description: >-
        Bouton « Infiltrer » de l'écran Établissements. Émet un jeton de MÊME format qu'un jeton
        normal (sub/schoolId/role, AGENTS.md règle #10), portant l'identité du Directeur actif de
        l'établissement ciblé, avec un claim `impersonatedBy` supplémentaire traçant l'acteur réel.
        AUCUN refresh token n'accompagne ce jeton : la session expire pour de bon avec l'access token
        (~15 min) et ne se renouvelle jamais — sortir plus tôt échange le cookie de refresh du Super
        Admin (resté intact) contre ses propres jetons. Écrit une entrée d'audit OBLIGATOIRE sur
        l'école ciblée avant de renvoyer le jeton.
      parameters:
        - name: schoolId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Jeton d'impersonation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImpersonateSchoolResult' }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
        '404': { description: Établissement introuvable }
        '409': { description: Aucun compte Directeur actif pour cet établissement }
  /admin/platform/schools/{schoolId}/complimentary-access:
    post:
      tags: [Platform]
      summary: Offrir un accès gratuit/complémentaire à un établissement (Super Admin uniquement)
      description: >-
        Aucun paiement : AGENTS.md règle #11 ne s'applique pas ici, ce n'est pas une confirmation
        d'encaissement. Écrit directement l'abonnement via la fonction SECURITY DEFINER
        `grant_complimentary_subscription` (le Super Admin n'a pas de SchoolId propre pour une écriture
        EF normale sur une table RLS). `durationMonths` entre 1 et 60.
      parameters:
        - name: schoolId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GrantComplimentaryAccessRequest' }
      responses:
        '204': { description: Accès complémentaire accordé }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
        '404': { description: Aucun abonnement associé à cet établissement }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/platform/schools/{schoolId}/sms-credits:
    post:
      tags: [Platform]
      summary: Créditer le solde SMS d'un établissement (Super Admin uniquement)
      description: >-
        Incrément atomique de `SchoolSettings.SmsCreditBalance` (en segments), hors RLS via
        `IgnoreQueryFilters` — le Super Admin n'a pas de tenant. `segments` entre 1 et 1 000 000.
      parameters:
        - name: schoolId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TopUpSmsCreditsRequest' }
      responses:
        '200':
          description: Nouveau solde
          content:
            application/json:
              schema:
                type: object
                properties:
                  creditBalance: { type: integer }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
        '404': { description: Aucun paramétrage trouvé pour cet établissement }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/platform/users/{userId}/schools:
    post:
      tags: [Platform]
      summary: Rattacher un utilisateur à un établissement supplémentaire (Super Admin uniquement)
      description: Idempotent — rattacher un établissement déjà associé répond 204 sans rien modifier.
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttachSchoolRequest' }
      responses:
        '204': { description: Établissement rattaché }
        '403': { description: Rôle appelant non autorisé (réservé au Super Admin) }
        '404': { description: Utilisateur ou établissement introuvable }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNumber:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
...
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Format normalisé — voir docs/Volume_4_API_Design.md §0.4
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
        traceId: { type: string }
...
    PlatformDashboardStats:
      type: object
      description: Ligne unique de la vue `v_platform_dashboard_stats` (GET /admin/platform/dashboard).
      properties:
        totalSchools: { type: integer }
        totalUsers: { type: integer }
        totalRevenue: { type: number, format: decimal, description: "Somme des paiements d'abonnement CONFIRMÉS uniquement (AGENTS.md règle #11)." }
        activeSubscriptions: { type: integer }
        mrr: { type: number, format: decimal, description: "Monthly Recurring Revenue — somme, sur les écoles à l'abonnement Active, du dernier paiement confirmé ramené à un équivalent mensuel (Yearly / 12)." }
        arr: { type: number, format: decimal, description: "Annual Recurring Revenue — MRR × 12." }
        forecastedRevenue30Days: { type: number, format: decimal, description: "Cashflow prévisionnel : somme des derniers paiements confirmés des écoles dont l'abonnement Active échoit dans les 30 prochains jours (ExpiresAt, pas un historique)." }
        arpu: { type: number, format: decimal, description: "Average Revenue Per User — MRR / nombre d'écoles à l'abonnement Active." }
    GlobalAuditLogItem:
      type: object
      description: Ligne de GET /admin/platform/activity — même nature que AuditLogEntry, avec l'établissement en plus.
      properties:
        id: { type: string, format: uuid }
        schoolId: { type: string, format: uuid }
        schoolName: { type: string }
        userId: { type: string, format: uuid }
        actorFullName: { type: string }
        module: { type: string, example: Finance }
        action: { type: string, example: RecordPayment }
        success: { type: boolean }
        failureReason: { type: string, nullable: true }
        ipAddress: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }
    PaginatedGlobalAuditLogs:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/GlobalAuditLogItem' } }
        totalCount: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    PlatformSubscription:
      type: object
      description: Ligne de GET /admin/platform/subscriptions — une école ayant un abonnement, avec son dernier paiement CONFIRMÉ le cas échéant.
      properties:
        schoolId: { type: string, format: uuid }
        schoolName: { type: string }
        plan: { type: string, enum: [Primaire, Standard, Premium] }
        status: { type: string, enum: [AwaitingPayment, Active, Suspended, ReadOnly] }
        expiresAt: { type: string, format: date, nullable: true }
        lastPaymentAmountXof: { type: number, format: decimal, nullable: true }
        lastPaymentAt: { type: string, format: date-time, nullable: true }
        lastPaymentBillingPeriod: { type: string, enum: [Monthly, Yearly], nullable: true, description: "Période de facturation du dernier paiement confirmé — alimente la colonne « Montant Contrat » de l'écran Abonnements & Facturation." }
    MonthlyRevenueProjection:
      type: object
      description: >-
        Un point de la courbe GET /admin/platform/revenue-projection : le cash attendu ce mois-là, basé
        sur les échéances RÉELLES des abonnements Active (pas une hypothèse de croissance).
      properties:
        month: { type: string, example: "2026-09", description: "Année-mois (yyyy-MM)." }
        projectedRevenue: { type: number, format: decimal }
    ImpersonateSchoolResult:
      type: object
      description: >-
        Même forme qu'AuthTokens à un champ près (expiresIn, pas de refresh token — voir la
        description de l'opération) : le client réutilise la même logique de session côté navigateur.
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, description: Durée de vie en secondes (~900, JwtOptions.AccessTokenMinutes). }
        schoolId: { type: string, format: uuid }
        schoolName: { type: string }
...
    GrantComplimentaryAccessRequest:
      type: object
      required: [plan, durationMonths]
      properties:
        plan: { type: string, enum: [Primaire, Standard, Premium] }
        durationMonths: { type: integer, minimum: 1, maximum: 60 }
    TopUpSmsCreditsRequest:
      type: object
      required: [segments]
      properties:
        segments: { type: integer, minimum: 1, maximum: 1000000 }
    AttachSchoolRequest:
      type: object
      required: [schoolId]
      properties:
        schoolId: { type: string, format: uuid }
...